fix: remove int32 casts, fix espnow fragment assert, drop redundant valid[] array
- tbft_state.c: remove stale (int32_t) casts in ptree_update_leaf calls
  now that tbft_part_t.version is tbft_seqno_t (int64_t)
- tbft_transport_espnow.c: fix frag_received comment (count, not bitmap),
  add _Static_assert(FRAG_MAX_PARTS <= 32) to guard the frag_mask uint32_t
- tbft_view_info.h: remove redundant valid[] array (duplicate of received[])
- tbft_view_info.c: replace all vi->valid[] references with vi->received[]

+++ src/tbft_state.c
@@ -268,8 +268,7 @@ void tbft_state_handle_data(tbft_state_t *state,
     memcpy(state->mem + (size_t)bidx * TBFT_BLOCK_SIZE,
            block_data, TBFT_BLOCK_SIZE);
     state->block_digests[bidx] = computed;
-    tbft_ptree_update_leaf(&state->ptree, bidx, &computed,
-                           (int32_t)rep->seqno);
+    tbft_ptree_update_leaf(&state->ptree, bidx, &computed, rep->seqno);
 
     /* Check if all pending fetches are resolved */
     bool all_done = true;
